Fix bug in iterating all libuditest testcases.
The test library was relying on the static initialization of
libuditest.o to occur before the static initialization of
the testcases. Since libuditest.o was being linked statically
with the testcases, there is no guaranteed order in which
static initialization will be called in each translation unit.

Ran into this bug while testing with a newer compiler, which
explains why I didn't see this sooner...

The fix is to move libuditest into its own .so because the
initialiation for a .so will be called before the initialization
of the executable, thereby providing the ordering required by
libuditest.