check result of malloc
There are a few other un-checked malloc()s here, but it's not obvious what the
convention for such things is. One convention I've used with moderate success
is to initialize the variables at the top (in true C style) to NULL, and then
malloc() and memset, and when something fucks up, goto out which then tries to
free everything that's not NULL. Anyway, here's one small easy fix.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>